Break time setting modal part into component
